feat: SessionStudent 저장 및 조회 추가

+++ src/main/java/nextstep/courses/domain/SessionRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain;
 import nextstep.courses.domain.session.Session;
 
 public interface SessionRepository {
-    int save(Session session);
+    Long save(Session session);
 
     Session findById(Long id);
 }
